pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent"
version = "0.1.0"
description = "一个支持多轮对话的AI Agent"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
]

[project.scripts]
agent = "agent.ai_agent:main"
ai-agent = "agent.ai_agent:main"

[project.urls]
Homepage = "https://github.com/yourusername/agent"
Repository = "https://github.com/yourusername/agent"

[tool.setuptools.packages.find]
where = ["src"]